+++ examples/trivial.rs
@@ -126,7 +126,7 @@ fn main_exploded() {
         flags::MEM_COPY_HOST_PTR), dims, Some(&vec)).unwrap();
 
     // (3) Create a kernel with arguments matching those in the source above:
-    let kernel = Kernel::new("add", &program, &queue).unwrap()
+    let kernel = Kernel::new("add", &program, queue.clone()).unwrap()
         .gws(&dims)
         .arg_buf(&buffer)
         .arg_scl(10.0f32);