+++ src/standard/pro_que.rs
@@ -306,7 +306,7 @@ impl ProQue {
 
     /// Creates a kernel with pre-assigned dimensions.
     pub fn create_kernel(&self, name: &str) -> OclResult<Kernel> {
-        let kernel = try!(Kernel::new(name.to_string(), &self.program, &self.queue));
+        let kernel = try!(Kernel::new(name.to_string(), &self.program, self.queue.clone()));
 
         match self.dims {
             Some(d) => Ok(kernel.gws(d)),